moved mapToNearestStation to the helper file for trains
